+++ Magic/CHANGELOG.md
@@ -41,6 +41,7 @@
  - Fix using the replicate and clone brushes with the Place spell
  - Fix pasting schematics with torches or other wall attachables
 
+ - The `/mconfig clean` command has been disabled (hopefully temporarily)
  - Add ignore_damage option for crafting recipes
  - Add auto_block_state for ModifyBlock, used by Place spell, attempts to place a block as if a player placed it
  - Add `/mwarp configure` command, used for Dynmap and Recall integration